Enabled passing in of host and port values via command line variables.
Created default host value of '127.0.0.1' and port value
of '8080'. Values passed in from the command line in the form
of key=value where key is host or port will override the
default values.

This functionality enables the configuration of the app without
having to checkout the code, modify the hard coded values, and
check the changes back in somewhere.